requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.2
langchain>=0.1.0
groq>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
